sync net/textproto #597 sync header
